Optimize getting best server algorithm.

+++ src/SpeedtestConfig.c
@@ -13,7 +13,7 @@
 
 const char *ConfigLineIdentitier[] = {"<client", "<upload", "<server-config", "<download"};
 
-long haversineDistance(float lat1, float lon1, float lat2, float lon2)
+float haversineDistance(float lat1, float lon1, float lat2, float lon2)
 {
     float dx, dy, dz, a, b;
     lon1 -= lon2;
